Added eight.png and four.png file. 	new file:   resources/eight.png 	new file:   resources/four.png
Modified A* Star Algorithm so that it can produce optimal results while
traversing in all eight directions as well.
	modified:   src/A_Star_Runnable.java

Modified AlgoRunnerRunnableImpl to decide whether it should produce
neighbours in all four directions or in all eight directions based upon
mapComp.isFourDir property, which the user can choose from UI
	modified:   src/AlgoRunnerRunnableImpl.java

Based upon the above change, following three programs has also been
modified to reflect the same.
	modified:   src/BestFirstSearchRunnable.java
	modified:   src/BreadthFirstSearchRunnable.java
	modified:   src/DepthFirstSearchRunnable.java

Added fourDir property to mapComp and added control to mainFrame so that
user can turn on or off this property
If fourDir is set to true: a cell will have maximum of four neighbours in North,
East, South and West direction
If it is turned off, it can have upto 8 neighbours in all possible
direction.
	modified:   src/MapComp.java
	modified:   src/mainFrame.java